Changed data type for entry_amount from double to decimal

+++ database/migrations/2020_03_07_201249_create_entry_table.php
@@ -21,7 +21,7 @@ class CreateEntryTable extends Migration
             $table->timestamp('created_at')->useCurrent();
             $table->timestamp('updated_at')->default(DB::raw('CURRENT_TIMESTAMP on update CURRENT_TIMESTAMP'));
             $table->string('entry_description');
-            $table->double('entry_amount');
+            $table->decimal('entry_amount', 12, 2);
         });
 
         Schema::table('entry', function($table){
